v1.0.0
Former-commit-id: d656d271d0e48040568165fb0d7a6c383dde7d17

+++ src/main/java/club/xiaojiawei/hearthstone/controller/DashboardController.java
@@ -3,21 +3,20 @@ package club.xiaojiawei.hearthstone.controller;
 
 import club.xiaojiawei.hearthstone.constant.SystemConst;
 import club.xiaojiawei.hearthstone.entity.Result;
+import club.xiaojiawei.hearthstone.entity.WsResult;
 import club.xiaojiawei.hearthstone.run.Core;
 import club.xiaojiawei.hearthstone.status.Deck;
-import club.xiaojiawei.hearthstone.utils.SystemUtil;
+import club.xiaojiawei.hearthstone.ws.WebSocketServer;
 import lombok.SneakyThrows;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
+import java.io.File;
 import java.io.FileOutputStream;
-import java.net.URL;
-import java.net.URLDecoder;
 import java.util.List;
+import java.util.Objects;
 
 import static club.xiaojiawei.hearthstone.constant.SystemConst.PROPERTIES;
 
@@ -38,6 +37,7 @@ public class DashboardController {
     @ResponseBody
     public void shutdown(){
         Core.setPause(true);
+        WebSocketServer.sendAllMessage(WsResult.ofScriptLog("系统关闭"));
         System.exit(0);
     }
 
@@ -46,6 +46,12 @@ public class DashboardController {
     public Result<Object> start(){
         Core.start();
         return Result.ofSuccess(null);
+    }
+
+    @RequestMapping("/getStatus")
+    @ResponseBody
+    public Result<Object> getStatus(){
+        return Result.ofSuccess(Core.getPause());
     }
 
     @RequestMapping("/getRunningTime")
@@ -64,22 +70,31 @@ public class DashboardController {
     @RequestMapping("/setGamePath")
     @ResponseBody
     public Result<Object> setGamePath(@RequestParam String path){
-        PROPERTIES.setProperty("gamepath", path);
-        try(FileOutputStream fileOutputStream = new FileOutputStream(SystemConst.getPath() + SystemConst.getName())){
-            PROPERTIES.store(fileOutputStream, "zerg");
+        if (new File(path).exists()){
+            if (!new File(path + "/" + SystemConst.GAME_PROGRAM_NAME).exists()){
+                return Result.ofFail("不是炉石路径，正确的路径下有" + SystemConst.GAME_PROGRAM_NAME + "文件，请重新填写");
+            }
+            PROPERTIES.setProperty("gamepath", path);
+            try(FileOutputStream fileOutputStream = new FileOutputStream(SystemConst.getPath() + SystemConst.getName())){
+                PROPERTIES.store(fileOutputStream, "zerg");
+            }
+            return Result.ofSuccess("重启后生效");
         }
-        return Result.ofSuccess(null);
+        return Result.ofFail("路径不存在，请核实后重新填写");
     }
 
     @SneakyThrows
     @RequestMapping("/setPlatformPath")
     @ResponseBody
     public Result<Object> setPlatformPath(@RequestParam String path){
-        PROPERTIES.setProperty("platformpath", path);
-        try(FileOutputStream fileOutputStream = new FileOutputStream(SystemConst.getPath() + SystemConst.getName())){
-            PROPERTIES.store(fileOutputStream, "zerg");
+        if (new File(path).exists()){
+            PROPERTIES.setProperty("platformpath", path);
+            try(FileOutputStream fileOutputStream = new FileOutputStream(SystemConst.getPath() + SystemConst.getName())){
+                PROPERTIES.store(fileOutputStream, "zerg");
+            }
+            return Result.ofSuccess("重启后生效");
         }
-        return Result.ofSuccess(null);
+        return Result.ofFail("路径不存在，请核实后重新填写");
     }
 
 }
